+++ evita/dom/os/process_unittest.cc
@@ -99,7 +99,8 @@ TEST_F(OsProcessTest, OsProcess_write_succeeded) {
   EXPECT_SCRIPT_TRUE("process instanceof Os.Process");
 
   std::vector<uint8_t> expected_bytes {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
-  mock_io_delegate()->SetCallResult("WriteFile", 0, expected_bytes.size());
+  mock_io_delegate()->SetCallResult("WriteFile", 0,
+                                    static_cast<int>(expected_bytes.size()));
 
   EXPECT_SCRIPT_VALID(
     "var transferred;"
